mingw makefiles changes to generate less warnings. iostream.h replaced with iostream

+++ libmoogutil/sio.cpp
@@ -27,12 +27,12 @@
 #include "sio.h"
 #include "rcsid.h"
 
-//RCSID("$Id: sio.cpp,v 1.1 2004/03/30 10:08:52 brainslayer Exp $");
+//RCSID("$Id: sio.cpp,v 1.2 2004/03/31 08:31:41 brainslayer Exp $");
 
 ssize_t readn(int fd, void *buf, size_t len)
 {
 
-  int nleft,nread;
+  size_t nleft,nread;
 
   nleft = len;
 
@@ -65,7 +65,7 @@ ssize_t readn(int fd, void *buf, size_t len)
 ssize_t writen(int fd, const void *buf, size_t len)
 {
   
-  int nleft, nwritten;
+  size_t nleft, nwritten;
 
   nleft = len;
